Fix type safety and code quality issues across codebase
- Fix narrowing conversion in C++ router: use size_t consistently for std::views::take
- Add dtype validation in storage model: restrict to float32/float64 only
- Remove inline numpy import in Python router: move to top-level imports
- Fix Python binding tests: add float64 router fixture for proper dtype testing

These changes improve type safety, prevent runtime errors, and ensure
code follows Python and C++ best practices for imports and validation.

+++ adaptive_router_core/core/include/adaptive_core/router.hpp
@@ -56,7 +56,7 @@ public:
     };
 
     if (scores.size() > 1) {
-      int n = std::min(static_cast<int>(scores.size()) - 1, profile_.metadata.routing.max_alternatives);
+      size_t n = std::min(scores.size() - 1, static_cast<size_t>(profile_.metadata.routing.max_alternatives));
       auto alts = scores | std::views::drop(1) | std::views::take(n)
                         | std::views::transform(&ModelScore::model_id);
       resp.alternatives.assign(alts.begin(), alts.end());
